finished in UI of invite students
